+++ src/components/app/app.jsx
@@ -1,7 +1,7 @@
-import Main from "@root/components/main/main";
-import {AppPages} from "@root/consts/main";
+import Main from "@components/main/main";
+import FilmPage from "@components/film-page/film-page";
+import {AppPages} from "@constants/main";
 import {BrowserRouter, Switch, Route} from "react-router-dom";
-import FilmPage from "@root/components/film-page/film-page";
 
 class App extends React.PureComponent {
   constructor(props) {
@@ -18,6 +18,7 @@ class App extends React.PureComponent {
   _renderApp() {
     const {promo, films} = this.props;
     const {page, film} = this.state;
+    const extraFilms = this._getExtraFilms(film, films);
 
     if (films) {
       switch (page) {
@@ -33,7 +34,7 @@ class App extends React.PureComponent {
           return (
             <FilmPage
               film={film}
-              films={films}
+              extraFilms={extraFilms}
               onFilmCardElementClick={this._onFilmCardElementClick}
             />
           );
@@ -41,6 +42,23 @@ class App extends React.PureComponent {
     }
 
     return null;
+  }
+
+  _getExtraFilms(film, films) {
+    const extraFilms = [];
+
+    if (film) {
+      for (const currentFilm of films) {
+        if (extraFilms.length === 4) {
+          break;
+        }
+        if (currentFilm.id !== film.id && currentFilm.genre === film.genre) {
+          extraFilms.push(currentFilm);
+        }
+      }
+    }
+
+    return extraFilms;
   }
 
   render() {
@@ -55,7 +73,7 @@ class App extends React.PureComponent {
           <Route exact path="/dev-details">
             <FilmPage
               film={films[1]}
-              films={films}
+              extraFilms={this._getExtraFilms(films[1], films)}
               onFilmCardElementClick={this._onFilmCardElementClick}
             />
           </Route>
