+++ src/components/film-page/film-page.jsx
@@ -1,24 +1,14 @@
-import FilmPageTabs from "@root/components/film-details-tabs/film-page-tabs";
-import FilmsList from "@root/components/films-list/films-list";
+import Tabs from "@components/tabs/tabs";
+import FilmsList from "@components/films-list/films-list";
 
 const FilmPage = (props) => {
-  const {film, films, onFilmCardElementClick} = props;
+  const {film, extraFilms, onFilmCardElementClick} = props;
   const {
     title,
     poster,
     releaseYear,
     genre,
   } = film;
-
-  const extraFilms = [];
-  for (const currentFilm of films) {
-    if (extraFilms.length === 4) {
-      break;
-    }
-    if (currentFilm !== film && currentFilm.genre === film.genre) {
-      extraFilms.push(currentFilm);
-    }
-  }
 
   return (
     <>
@@ -81,7 +71,7 @@ const FilmPage = (props) => {
             </div>
 
             <div className="movie-card__desc">
-              <FilmPageTabs film={film}/>
+              <Tabs film={film}/>
             </div>
           </div>
         </div>
@@ -118,7 +108,7 @@ const FilmPage = (props) => {
 };
 
 FilmPage.propTypes = {
-  films: PropTypes.arrayOf(PropTypes.shape({
+  extraFilms: PropTypes.arrayOf(PropTypes.shape({
     id: PropTypes.string.isRequired,
     title: PropTypes.string.isRequired,
     poster: PropTypes.string.isRequired,
