+++ src/index.jsx
@@ -1,9 +1,7 @@
 import ReactDOM from "react-dom";
-import App from "@root/components/app/app";
-import {MOCK_FILMS_COUNT, PromoSetting} from "@root/consts/main";
+import App from "@components/app/app";
+import {MOCK_FILMS_COUNT, PromoSetting} from "@constants/main";
 import {generateFilms} from "@root/mocks/films";
-
-console.log(generateFilms(MOCK_FILMS_COUNT));
 
 ReactDOM.render(
     <App
