+++ webpack.config.js
@@ -38,6 +38,9 @@ module.exports = (env) => {
     resolve: {
       alias: {
         '@root': path.resolve(__dirname, './src'),
+        '@components': path.resolve(__dirname, './src/components'),
+        '@utils': path.resolve(__dirname, './src/utils'),
+        '@constants': path.resolve(__dirname, './src/constants'),
       },
       extensions: ['*', '.js', '.jsx']
     }
